move pages into own folders
