+++ src/pages/index.js
@@ -1,3 +1,3 @@
-export * from './Page1'
-export * from './Page2'
-export * from './Page3'
+export * from './Page1/Page1'
+export * from './Page2/Page2'
+export * from './Page3/Page3'
